fix: Derive variable types from IR expressions, not TypeScript
For module-level variables without explicit type annotations, the type
is now derived from the converted IR expression using deterministic rules:

- Integer literals (numericIntent: Int32) → int
- Floating literals (numericIntent: Double) → double
- String literals → string
- Boolean literals → boolean
- Arrays → element type derived from first element

This eliminates TypeScript type inference from type determination,
ensuring `const value = 100` produces `int value = 100` in C#,
not `double value = 100`.

Updated expected files for constants and variable-decls golden tests.

+++ packages/emitter/testcases/common/expected/types/constants/ModuleConstants.cs
@@ -6,7 +6,7 @@ namespace TestCases.common.types.constants
 
             public static readonly string MESSAGE = "Hello, World!";
 
-            public static readonly double COUNT = 42;
+            public static readonly int COUNT = 42;
 
             public static readonly bool IS_ENABLED = true;
         }
